com.unity.services.ccd.management@3.0.0 ## [3.0.0] - 2023-11-27 ### Added * Added support for ListBucketsOptions for the ListBucketsAsync method. * Added RoleOptions to DeletePermissionAsync method.
### Changed
* Fix integer overflow when ContentSize is larger than 2GB by changing ContentSize to a long.
* Aligned most integer types to use longs instead of ints to properly reflect the API.
* Fix issue with CcdSampleWindow.cs sample causing errors in standalone build.
* Updated Core and NewtonsoftJson dependencies.

## [2.3.0] - 2023-12-15
### Added
* Added LogRequests and LogRequestHeaders to CcdManagement to add a way to log internal requests
### Changed
* Catch 401 and 403 errors, get new token, and retry

+++ Editor/Models/CcdBadge.cs
@@ -38,7 +38,7 @@ namespace Unity.Services.Ccd.Management.Models
         /// <param name="releaseid">releaseid param</param>
         /// <param name="releasenum">releasenum param</param>
         [Preserve]
-        public CcdBadge(DateTime created = default, string createdBy = default, string createdByName = default, string name = default, System.Guid releaseid = default, int releasenum = default)
+        public CcdBadge(DateTime created = default, string createdBy = default, string createdByName = default, string name = default, System.Guid releaseid = default, long releasenum = default)
         {
             Created = created;
             CreatedBy = createdBy;
@@ -88,7 +88,7 @@ namespace Unity.Services.Ccd.Management.Models
         /// </summary>
         [Preserve]
         [DataMember(Name = "releasenum", EmitDefaultValue = false)]
-        public int Releasenum{ get; }
+        public long Releasenum{ get; }
     
         /// <summary>
         /// Formats a CcdBadge into a string of key-value pairs for use as a path parameter.
